Create Guess the number game
A number guessing game ,user need to guess a no between 1-10, ctime liabrary is used in this function to seed the random number generator with a different seed value each time you run the program,cstdlib  is used to seed the random number generator before using std::rand(). It's also declared in the cstdlib header.It also tells user that the number that is generated is lower or higher than the number that is guessed.